project/03/sd_vae_lightning/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
pytorch-lightning>=2.0.0
diffusers>=0.25.0
transformers>=4.30.0
datasets>=2.14.0
huggingface-hub>=0.16.0
lpips>=0.1.4
tensorboard>=2.14.0
Pillow>=9.0.0
numpy>=1.24.0
tqdm>=4.65.0
accelerate>=0.20.0
einops>=0.6.0
omegaconf>=2.3.0
torchmetrics>=1.0.0
scikit-image>=0.21.0